BAU: fix fixture score for AWS integration tests
...
